refactor(graph_engine): move observability layer and persistence layer to core.app.workflow
