+++ api/core/app/workflow/layers/__init__.py
@@ -1,0 +1,10 @@
+"""Workflow-level GraphEngine layers that depend on outer infrastructure."""
+
+from .observability import ObservabilityLayer
+from .persistence import PersistenceWorkflowInfo, WorkflowPersistenceLayer
+
+__all__ = [
+    "ObservabilityLayer",
+    "PersistenceWorkflowInfo",
+    "WorkflowPersistenceLayer",
+]
