+++ api/tests/unit_tests/core/workflow/graph_engine/layers/test_observability.py
@@ -15,14 +15,14 @@ from unittest.mock import patch
 import pytest
 from opentelemetry.trace import StatusCode
 
+from core.app.workflow.layers.observability import ObservabilityLayer
 from core.workflow.enums import NodeType
-from core.workflow.graph_engine.layers.observability import ObservabilityLayer
 
 
 class TestObservabilityLayerInitialization:
     """Test ObservabilityLayer initialization logic."""
 
-    @patch("core.workflow.graph_engine.layers.observability.dify_config.ENABLE_OTEL", True)
+    @patch("core.app.workflow.layers.observability.dify_config.ENABLE_OTEL", True)
     @pytest.mark.usefixtures("mock_is_instrument_flag_enabled_false")
     def test_initialization_when_otel_enabled(self, tracer_provider_with_memory_exporter):
         """Test that layer initializes correctly when OTel is enabled."""
@@ -32,7 +32,7 @@ class TestObservabilityLayerInitialization:
         assert NodeType.TOOL in layer._parsers
         assert layer._default_parser is not None
 
-    @patch("core.workflow.graph_engine.layers.observability.dify_config.ENABLE_OTEL", False)
+    @patch("core.app.workflow.layers.observability.dify_config.ENABLE_OTEL", False)
     @pytest.mark.usefixtures("mock_is_instrument_flag_enabled_true")
     def test_initialization_when_instrument_flag_enabled(self, tracer_provider_with_memory_exporter):
         """Test that layer enables when instrument flag is enabled."""
@@ -46,7 +46,7 @@ class TestObservabilityLayerInitialization:
 class TestObservabilityLayerNodeSpanLifecycle:
     """Test node span creation and lifecycle management."""
 
-    @patch("core.workflow.graph_engine.layers.observability.dify_config.ENABLE_OTEL", True)
+    @patch("core.app.workflow.layers.observability.dify_config.ENABLE_OTEL", True)
     @pytest.mark.usefixtures("mock_is_instrument_flag_enabled_false")
     def test_node_span_created_and_ended(
         self, tracer_provider_with_memory_exporter, memory_span_exporter, mock_llm_node
@@ -63,7 +63,7 @@ class TestObservabilityLayerNodeSpanLifecycle:
         assert spans[0].name == mock_llm_node.title
         assert spans[0].status.status_code == StatusCode.OK
 
-    @patch("core.workflow.graph_engine.layers.observability.dify_config.ENABLE_OTEL", True)
+    @patch("core.app.workflow.layers.observability.dify_config.ENABLE_OTEL", True)
     @pytest.mark.usefixtures("mock_is_instrument_flag_enabled_false")
     def test_node_error_recorded_in_span(
         self, tracer_provider_with_memory_exporter, memory_span_exporter, mock_llm_node
@@ -82,7 +82,7 @@ class TestObservabilityLayerNodeSpanLifecycle:
         assert len(spans[0].events) > 0
         assert any("exception" in event.name.lower() for event in spans[0].events)
 
-    @patch("core.workflow.graph_engine.layers.observability.dify_config.ENABLE_OTEL", True)
+    @patch("core.app.workflow.layers.observability.dify_config.ENABLE_OTEL", True)
     @pytest.mark.usefixtures("mock_is_instrument_flag_enabled_false")
     def test_node_end_without_start_handled_gracefully(
         self, tracer_provider_with_memory_exporter, memory_span_exporter, mock_llm_node
@@ -100,7 +100,7 @@ class TestObservabilityLayerNodeSpanLifecycle:
 class TestObservabilityLayerParserIntegration:
     """Test parser integration for different node types."""
 
-    @patch("core.workflow.graph_engine.layers.observability.dify_config.ENABLE_OTEL", True)
+    @patch("core.app.workflow.layers.observability.dify_config.ENABLE_OTEL", True)
     @pytest.mark.usefixtures("mock_is_instrument_flag_enabled_false")
     def test_default_parser_used_for_regular_node(
         self, tracer_provider_with_memory_exporter, memory_span_exporter, mock_start_node
@@ -119,7 +119,7 @@ class TestObservabilityLayerParserIntegration:
         assert attrs["node.execution_id"] == mock_start_node.execution_id
         assert attrs["node.type"] == mock_start_node.node_type.value
 
-    @patch("core.workflow.graph_engine.layers.observability.dify_config.ENABLE_OTEL", True)
+    @patch("core.app.workflow.layers.observability.dify_config.ENABLE_OTEL", True)
     @pytest.mark.usefixtures("mock_is_instrument_flag_enabled_false")
     def test_tool_parser_used_for_tool_node(
         self, tracer_provider_with_memory_exporter, memory_span_exporter, mock_tool_node
@@ -138,7 +138,7 @@ class TestObservabilityLayerParserIntegration:
         assert attrs["gen_ai.tool.name"] == mock_tool_node.title
         assert attrs["gen_ai.tool.type"] == mock_tool_node._node_data.provider_type.value
 
-    @patch("core.workflow.graph_engine.layers.observability.dify_config.ENABLE_OTEL", True)
+    @patch("core.app.workflow.layers.observability.dify_config.ENABLE_OTEL", True)
     @pytest.mark.usefixtures("mock_is_instrument_flag_enabled_false")
     def test_llm_parser_used_for_llm_node(
         self, tracer_provider_with_memory_exporter, memory_span_exporter, mock_llm_node, mock_result_event
@@ -176,7 +176,7 @@ class TestObservabilityLayerParserIntegration:
         assert attrs["gen_ai.completion"] == "test completion"
         assert attrs["gen_ai.response.finish_reason"] == "stop"
 
-    @patch("core.workflow.graph_engine.layers.observability.dify_config.ENABLE_OTEL", True)
+    @patch("core.app.workflow.layers.observability.dify_config.ENABLE_OTEL", True)
     @pytest.mark.usefixtures("mock_is_instrument_flag_enabled_false")
     def test_retrieval_parser_used_for_retrieval_node(
         self, tracer_provider_with_memory_exporter, memory_span_exporter, mock_retrieval_node, mock_result_event
@@ -204,7 +204,7 @@ class TestObservabilityLayerParserIntegration:
         assert attrs["retrieval.query"] == "test query"
         assert "retrieval.document" in attrs
 
-    @patch("core.workflow.graph_engine.layers.observability.dify_config.ENABLE_OTEL", True)
+    @patch("core.app.workflow.layers.observability.dify_config.ENABLE_OTEL", True)
     @pytest.mark.usefixtures("mock_is_instrument_flag_enabled_false")
     def test_result_event_extracts_inputs_and_outputs(
         self, tracer_provider_with_memory_exporter, memory_span_exporter, mock_start_node, mock_result_event
@@ -235,7 +235,7 @@ class TestObservabilityLayerParserIntegration:
 class TestObservabilityLayerGraphLifecycle:
     """Test graph lifecycle management."""
 
-    @patch("core.workflow.graph_engine.layers.observability.dify_config.ENABLE_OTEL", True)
+    @patch("core.app.workflow.layers.observability.dify_config.ENABLE_OTEL", True)
     @pytest.mark.usefixtures("mock_is_instrument_flag_enabled_false")
     def test_on_graph_start_clears_contexts(self, tracer_provider_with_memory_exporter, mock_llm_node):
         """Test that on_graph_start clears node contexts."""
@@ -248,7 +248,7 @@ class TestObservabilityLayerGraphLifecycle:
         layer.on_graph_start()
         assert len(layer._node_contexts) == 0
 
-    @patch("core.workflow.graph_engine.layers.observability.dify_config.ENABLE_OTEL", True)
+    @patch("core.app.workflow.layers.observability.dify_config.ENABLE_OTEL", True)
     @pytest.mark.usefixtures("mock_is_instrument_flag_enabled_false")
     def test_on_graph_end_with_no_unfinished_spans(
         self, tracer_provider_with_memory_exporter, memory_span_exporter, mock_llm_node
@@ -264,7 +264,7 @@ class TestObservabilityLayerGraphLifecycle:
         spans = memory_span_exporter.get_finished_spans()
         assert len(spans) == 1
 
-    @patch("core.workflow.graph_engine.layers.observability.dify_config.ENABLE_OTEL", True)
+    @patch("core.app.workflow.layers.observability.dify_config.ENABLE_OTEL", True)
     @pytest.mark.usefixtures("mock_is_instrument_flag_enabled_false")
     def test_on_graph_end_with_unfinished_spans_logs_warning(
         self, tracer_provider_with_memory_exporter, mock_llm_node, caplog
@@ -285,7 +285,7 @@ class TestObservabilityLayerGraphLifecycle:
 class TestObservabilityLayerDisabledMode:
     """Test behavior when layer is disabled."""
 
-    @patch("core.workflow.graph_engine.layers.observability.dify_config.ENABLE_OTEL", False)
+    @patch("core.app.workflow.layers.observability.dify_config.ENABLE_OTEL", False)
     @pytest.mark.usefixtures("mock_is_instrument_flag_enabled_false")
     def test_disabled_mode_skips_node_start(self, memory_span_exporter, mock_start_node):
         """Test that disabled layer doesn't create spans on node start."""
@@ -299,7 +299,7 @@ class TestObservabilityLayerDisabledMode:
         spans = memory_span_exporter.get_finished_spans()
         assert len(spans) == 0
 
-    @patch("core.workflow.graph_engine.layers.observability.dify_config.ENABLE_OTEL", False)
+    @patch("core.app.workflow.layers.observability.dify_config.ENABLE_OTEL", False)
     @pytest.mark.usefixtures("mock_is_instrument_flag_enabled_false")
     def test_disabled_mode_skips_node_end(self, memory_span_exporter, mock_llm_node):
         """Test that disabled layer doesn't process node end."""
